refactor: export raw
