+++ .changeset/clean-rabbits-tickle.md
@@ -2,5 +2,5 @@
 "@vue-macros/jsx-directive": patch
 ---
 
-expose plugin to support browser environments
+expose raw to support browser environments
