+++ packages/jsx-directive/src/raw.ts
@@ -1,0 +1,3 @@
+import { plugin } from './core/plugin'
+
+export default plugin
